App/requirements.txt
python-dotenv==0.19.1
streamlit==1.38.0
google-generativeai
